requirements.txt
streamlit
python-dotenv
google-generativeai
Pillow
audio-recorder-streamlit
SpeechRecognition
pydub
